requirements.txt
pandas
google-cloud-bigquery
python-dotenv
